Update SinglyLinkedList
When sending value to add or delete there was no information that if the value is not present in the linked list. So I have added a print statement that will give an information that when the value is not present in the linked list. 

I also added conditions which checks  the remove_by_value function if the value is present in the head, then there is no need in iterating the linked list.